+++ kf_lib/actors/fighter/_quotes.py
@@ -3,6 +3,7 @@ from abc import ABC
 
 from kf_lib.actors.fighter._abc import FighterAPI
 from kf_lib.actors import quotes
+from kf_lib.ui import dim
 
 
 class QuoteMethods(FighterAPI, ABC):
@@ -16,7 +17,7 @@ class QuoteMethods(FighterAPI, ABC):
         pool = quotes.PREFIGHT_QUOTES.get(self.quotes, None)
         if pool is not None:
             q = random.choice(pool)
-            self.current_fight.show(f'{self.name}: "{q}"')
+            self.current_fight.show(dim(f'{self.name}: "{q}"'))
             return True
         else:  # this is important for correctly making pauses after quotes
             return False
@@ -25,4 +26,4 @@ class QuoteMethods(FighterAPI, ABC):
         pool = quotes.WIN_QUOTES.get(self.quotes, None)
         if pool is not None:
             q = random.choice(pool)
-            self.current_fight.show(f'{self.name}: "{q}"')
+            self.current_fight.show(dim(f'{self.name}: "{q}"'))